Move deployment can into deployment model.
