+++ src/models/index.ts
@@ -17,7 +17,7 @@ export * from './CronStringLengthError'
 export * from './Deployment'
 export * from './DeploymentFlowRunCreate'
 export * from './DeploymentUpdate'
-export * from './DeploymentWithObjectLevelScopes'
+export * from './DeploymentCan'
 export * from './EmpiricalPolicy'
 export * from './ExistingHandleError'
 export * from './FilterDateError'
